refactored, added some setting for webpack and ts configs, added absolute import
