+++ docs/src/components/pages/Tokens/TokensHeader.tsx
@@ -3,7 +3,7 @@ import {Link, Title, useAdaptivity} from '@vkontakte/vkui';
 import clsx from 'clsx';
 import React, {FC} from 'react';
 
-import packageJson from '../../../../../package.json';
+import packageJson from '@/../../package.json';
 
 const version = {
 	href: 'https://github.com/VKCOM/vkui-tokens/releases',
