+++ docs/webpack/webpack.config.js
@@ -1,30 +1,17 @@
-// const path = require('path');
+const path = require('path');
 
 module.exports = function () {
 	const modules = {
 		ts: {
 			test: /\.ts(x?)$/,
-			use: 'ts-loader',
-			exclude: /node_modules/,
-		},
-		js: {
-			test: /\.(js)$/,
 			use: [
 				{
 					loader: 'babel-loader',
-					options: {
-						presets: [
-							[
-								'@babel/preset-env',
-								{
-									useBuiltIns: 'entry',
-								},
-							],
-						],
-					},
+				},
+				{
+					loader: 'ts-loader',
 				},
 			],
-			exclude: /node_modules/,
 		},
 		css: {
 			test: /\.css$/,
@@ -49,10 +36,10 @@ module.exports = function () {
 
 	const resolve = {
 		extensions: ['.ts', '.tsx', '.js', '.jsx'],
-		// alias: {
-		// 	'@/*': path.resolve(__dirname, '../src/*'),
-		// 	'@/public': path.resolve(__dirname, '../public/*'),
-		// },
+		alias: {
+			'@/public': path.resolve(__dirname, '../public'),
+			'@': path.resolve(__dirname, '../src'),
+		},
 	};
 
 	return {
