+++ package.json
@@ -29,11 +29,11 @@
   },
   "repository": "https://github.com/VKCOM/vkui-tokens.git",
   "dependencies": {
+    "core-js": "3.0.0",
     "csstype": "3.1.1"
   },
   "devDependencies": {
     "@babel/core": "7.19.3",
-    "@babel/preset-env": "7.19.4",
     "@svgr/webpack": "6.4.0",
     "@types/color": "3.0.1",
     "@types/common-tags": "1.8.1",
